+++ router.js
@@ -275,7 +275,7 @@ if (config.github && config.github.access_token) {
 }
 router.get('/github/trending/:since/:language?', require('./routes/github/trending'));
 router.get('/github/issue/:user/:repo', require('./routes/github/issue'));
-router.get('/github/user/:user', require('./routes/github/user'));
+router.get('/github/user/follower/:user', require('./routes/github/follower'));
 
 // konachan
 router.get('/konachan/post/popular_recent', require('./routes/konachan/post_popular_recent'));
